Move LinkFragment into Core/AppLink, resolve links from __typename in AppLink
