+++ src/components/Blocks/ImageCardGrid/ImageCardGrid.vue
@@ -34,7 +34,7 @@
           <AppLink
             v-if="card.link"
             class="image-card-grid__link app-button body app-button--secondary"
-            :to="card.link.to"
+            :link="card.link.record"
           >
             {{ card.link.label }} &rarr;
           </AppLink>
@@ -47,9 +47,9 @@
 <script setup lang="ts">
 import { type ImageCardGridFragment } from "./ImageCardGrid.query";
 import { type FragmentOf, readFragment } from "~/utils/graphql";
-import { resolveLink } from "~/utils/link";
 import { BackgroundColor } from "~/types/styling";
 import AppLink from "~/components/Core/AppLink/AppLink.vue";
+import { resolveLink } from "~/utils/link";
 
 const props = defineProps<{
   data: FragmentOf<typeof ImageCardGridFragment>;
@@ -60,7 +60,7 @@ const data = readFragment<typeof ImageCardGridFragment>(props.data);
 const cards = computed(() => {
   return data.items.map((card) => ({
     ...card,
-    link: card.links[0] ? resolveLink(card.links[0]) : undefined,
+    link: card.links[0] ? { record: card.links[0], ...resolveLink(card.links[0]) } : undefined,
   }));
 });
